[Services] organize structure of project Create models & services storage
