+++ src/Application.java
@@ -1,0 +1,15 @@
+
+public class Application {
+
+    public static void main(String[] args) {
+        //Create MessageStorageService
+
+        //Write program for query & save message here
+
+        //List menu here
+
+        //Using scanner to get action of user
+
+        //Perform action following by input
+    }
+}
