+++ src/com/chat/services/MessageStorageService.java
@@ -1,0 +1,32 @@
+package com.chat.services;
+
+import com.chat.models.Conversation;
+
+public class MessageStorageService {
+
+
+    public void saveConversation(Conversation conversation) {
+        //Save conversation to file with file name: conversation id
+
+        //Create file write with conversation id
+
+        //Loop message of conversation & write
+
+        //Close file
+    }
+
+    public Conversation getConversation() {
+        //Read file & create conversation
+        Conversation conversation = null;
+
+        //Create file reader with conversation id
+
+        //Read line by line
+
+        //Add to list of conversation
+
+
+
+        return conversation;
+    }
+}
